worker number adjust to the P number(if specified)
