+++ pkg/config/parser_test.go
@@ -43,6 +43,7 @@ var cb testCallback
 func TestMain(m *testing.M) {
 	RegisterConfigParsedListener(ParseCallbackKeyCluster, cb.ParsedCallback)
 	RegisterConfigParsedListener(ParseCallbackKeyServiceRgtInfo, cb.ParsedCallback)
+	RegisterConfigParsedListener(ParseCallbackKeyProcessor, cb.ParsedCallback)
 	m.Run()
 }
