+++ pkg/proxy/proxy.go
@@ -33,6 +33,7 @@ import (
 	"github.com/alipay/sofa-mosn/pkg/types"
 	"github.com/json-iterator/go"
 	"github.com/rcrowley/go-metrics"
+	"github.com/alipay/sofa-mosn/pkg/config"
 )
 
 var json = jsoniter.ConfigCompatibleWithStandardLibrary
@@ -47,13 +48,24 @@ var (
 func init() {
 	globalStats = newProxyStats(types.GlobalProxyName)
 
+	// register init function with interest of P number
+	config.RegisterConfigParsedListener(config.ParseCallbackKeyProcessor, initWorkePpool)
+}
+
+func initWorkePpool(data interface{}, endParsing bool) error {
 	// default shardsNum is equal to the cpu num
 	shardsNum := runtime.NumCPU()
 	// use 4096 as chan buffer length
 	poolSize := shardsNum * 4096
 
+	// set shardsNum equal to processor if it was specified
+	if pNum, ok := data.(int); ok && pNum > 0 {
+		shardsNum = pNum
+	}
+
 	workerPool, _ = mosnsync.NewShardWorkerPool(poolSize, shardsNum, eventDispatch)
 	workerPool.Init()
+	return nil
 }
 
 // types.ReadFilter
