+++ pkg/proxy/types.go
@@ -71,16 +71,3 @@ const (
 	UpstreamGlobalTimeout UpstreamResetType = "UpstreamGlobalTimeout"
 	UpstreamPerTryTimeout UpstreamResetType = "UpstreamPerTryTimeout"
 )
-
-func init() {
-	ConnNewPoolFactories = make(map[types.Protocol]connNewPool)
-}
-
-type connNewPool func(host types.Host) types.ConnectionPool
-
-var ConnNewPoolFactories map[types.Protocol]connNewPool
-
-func RegisterNewPoolFactory(protocol types.Protocol, factory connNewPool) {
-	//other
-	ConnNewPoolFactories[protocol] = factory
-}
