+++ pkg/stream/xprotocol/connpool.go
@@ -25,14 +25,14 @@ import (
 
 	"github.com/alipay/sofa-mosn/pkg/log"
 	"github.com/alipay/sofa-mosn/pkg/protocol"
-	"github.com/alipay/sofa-mosn/pkg/proxy"
 	str "github.com/alipay/sofa-mosn/pkg/stream"
 	"github.com/alipay/sofa-mosn/pkg/types"
 	"github.com/rcrowley/go-metrics"
+	"github.com/alipay/sofa-mosn/pkg/network"
 )
 
 func init() {
-	proxy.RegisterNewPoolFactory(protocol.Xprotocol, NewConnPool)
+	network.RegisterNewPoolFactory(protocol.Xprotocol, NewConnPool)
 	types.RegisterConnPoolFactory(protocol.Xprotocol, true)
 
 }
